backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
litellm>=1.60.0
pydantic>=2.10.0
typer>=0.15.0
rich>=13.9.0
yfinance>=0.2.50
python-dotenv>=1.0.0
